fix(clustering): prevent overflow in ARI expected index, X-means variance and G-means correction
CodeQL cs/loss-of-precision: AdjustedRandIndex.cs:142 multiplied two long pair counts
(overflows long at ~200k points; ARI of independent labelings came out 0.48 instead of ~0);
XMeans.cs:404 points*dims and GMeans.cs:363 n*n were int products.

+++ src/Clustering/AutoK/GMeans.cs
@@ -360,7 +360,7 @@ public partial class GMeans<T> : ClusteringBase<T>
         double ad = -n - sum / n;
 
         // Apply correction for small samples
-        ad *= (1 + 4.0 / n - 25.0 / (n * n));
+        ad *= (1 + 4.0 / n - 25.0 / ((double)n * n));
 
         return ad;
     }
